pyproject.toml
[project]
name = "agrollm-hack"
version = "0.1.0"
description = ""
authors = [
    {name = "UllubiyZabitov",email = "unzabitov@edu.hse.ru"}
]
package-mode = false
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pydantic (>=2.11.3,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "selenium (>=4.31.0,<5.0.0)",
    "streamlit (>=1.44.1,<2.0.0)",
    "webdriver-manager (>=4.0.2,<5.0.0)",
    "apscheduler (>=3.11.0,<4.0.0)",
    "plotly (>=6.0.1,<7.0.0)",
    "streamlit-autorefresh (>=1.0.1,<2.0.0)",
    "psutil (>=7.0.0,<8.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

